requirements.txt
# DO NOT include azure-functions-worker in this file
# The Python Worker is managed by Azure Functions platform
# Manually managing azure-functions-worker may cause unexpected issues

azure-functions
azure-functions-durable
attrs==23.1.0
Automat==22.10.0
azure-common==1.1.28
azure-core==1.27.0
azure-identity==1.13.0
azure-keyvault-secrets==4.7.0
azure-storage-blob==12.16.0
certifi==2023.5.7
cffi==1.15.1
charset-normalizer==3.1.0
colorama==0.4.6
constantly==15.1.0
cryptography==40.0.2
cssselect==1.2.0
dateutils==0.6.12
exceptiongroup==1.1.3
filelock==3.12.0
hyperlink==21.0.0
idna==3.4
incremental==22.10.0
iniconfig==2.0.0
isodate==0.6.1
itemadapter==0.8.0
itemloaders==1.1.0
jmespath==1.0.1
lxml==4.9.2
msal==1.22.0
msal-extensions==1.0.0
openpyxl==3.1.2
packaging==23.1
pandas==2.1.1
parsel==1.8.1
pluggy==1.3.0
portalocker==2.7.0
Protego==0.2.1
pyasn1==0.5.0
pyasn1-modules==0.3.0
pycparser==2.21
pyodbc===5.1.0
PyDispatcher==2.0.7
PyJWT==2.7.0
pyOpenSSL==23.1.1
pytest==7.4.2
pytest-azurepipelines==1.0.4
pytest-nunit==1.0.3
python-dateutil==2.8.2
python-dotenv==1.0.0
pytz==2023.3
queuelib==1.6.2
requests==2.30.0
requests-file==1.5.1
Scrapy==2.9.0
service-identity==21.1.0
six==1.16.0
tldextract==3.4.1
tomli==2.0.1
Twisted==22.10.0
typing_extensions==4.5.0
urllib3==2.0.2
w3lib==2.1.1
zope.interface==6.0
